Initial structure for new pdata metrics exporter
