+++ exporter/collector/googlecloud.go
@@ -21,20 +21,14 @@ import (
 	"fmt"
 	"strings"
 
-	"contrib.go.opencensus.io/exporter/stackdriver"
-	agentmetricspb "github.com/census-instrumentation/opencensus-proto/gen-go/agent/metrics/v1"
-	metricspb "github.com/census-instrumentation/opencensus-proto/gen-go/metrics/v1"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
 	"go.opentelemetry.io/collector/model/pdata"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"google.golang.org/api/option"
 	"google.golang.org/grpc"
 
 	cloudtrace "github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/trace"
-
-	internaldata "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/opencensus"
 )
 
 // traceExporter is a wrapper struct of OT cloud trace exporter
@@ -42,19 +36,8 @@ type traceExporter struct {
 	texporter *cloudtrace.Exporter
 }
 
-// metricsExporter is a wrapper struct of OC stackdriver exporter
-type metricsExporter struct {
-	mexporter *stackdriver.Exporter
-}
-
 func (te *traceExporter) Shutdown(ctx context.Context) error {
 	return te.texporter.Shutdown(ctx)
-}
-
-func (me *metricsExporter) Shutdown(context.Context) error {
-	me.mexporter.Flush()
-	me.mexporter.StopMetricsExporter()
-	return me.mexporter.Close()
 }
 
 func setVersionInUserAgent(cfg *Config, version string) {
@@ -123,117 +106,6 @@ func newGoogleCloudTracesExporter(cfg *Config, set component.ExporterCreateSetti
 		exporterhelper.WithRetry(cfg.RetrySettings))
 }
 
-func newGoogleCloudMetricsExporter(cfg *Config, set component.ExporterCreateSettings) (component.MetricsExporter, error) {
-	setVersionInUserAgent(cfg, set.BuildInfo.Version)
-
-	// TODO:  For each ProjectID, create a different exporter
-	// or at least a unique Google Cloud client per ProjectID.
-	options := stackdriver.Options{
-		// If the project ID is an empty string, it will be set by default based on
-		// the project this is running on in GCP.
-		ProjectID: cfg.ProjectID,
-
-		MetricPrefix: cfg.MetricConfig.Prefix,
-
-		// Set DefaultMonitoringLabels to an empty map to avoid getting the "opencensus_task" label
-		DefaultMonitoringLabels: &stackdriver.Labels{},
-
-		Timeout: cfg.Timeout,
-	}
-
-	// note options.UserAgent overrides the option.WithUserAgent client option in the Metric exporter
-	if cfg.UserAgent != "" {
-		options.UserAgent = cfg.UserAgent
-	}
-
-	copts, err := generateClientOptions(cfg)
-	if err != nil {
-		return nil, err
-	}
-	options.TraceClientOptions = copts
-	options.MonitoringClientOptions = copts
-
-	if cfg.MetricConfig.SkipCreateMetricDescriptor {
-		options.SkipCMD = true
-	}
-	if len(cfg.ResourceMappings) > 0 {
-		rm := resourceMapper{
-			mappings: cfg.ResourceMappings,
-		}
-		options.MapResource = rm.mapResource
-	}
-
-	sde, serr := stackdriver.NewExporter(options)
-	if serr != nil {
-		return nil, fmt.Errorf("cannot configure Google Cloud metric exporter: %w", serr)
-	}
-	mExp := &metricsExporter{mexporter: sde}
-
-	return exporterhelper.NewMetricsExporter(
-		cfg,
-		set,
-		mExp.pushMetrics,
-		exporterhelper.WithShutdown(mExp.Shutdown),
-		// Disable exporterhelper Timeout, since we are using a custom mechanism
-		// within exporter itself
-		exporterhelper.WithTimeout(exporterhelper.TimeoutSettings{Timeout: 0}),
-		exporterhelper.WithQueue(cfg.QueueSettings),
-		exporterhelper.WithRetry(cfg.RetrySettings))
-}
-
-// pushMetrics calls StackdriverExporter.PushMetricsProto on each element of the given metrics
-func (me *metricsExporter) pushMetrics(ctx context.Context, m pdata.Metrics) error {
-	rms := m.ResourceMetrics()
-	mds := make([]*agentmetricspb.ExportMetricsServiceRequest, 0, rms.Len())
-	for i := 0; i < rms.Len(); i++ {
-		emsr := &agentmetricspb.ExportMetricsServiceRequest{}
-		emsr.Node, emsr.Resource, emsr.Metrics = internaldata.ResourceMetricsToOC(rms.At(i))
-		mds = append(mds, emsr)
-	}
-	// PushMetricsProto doesn't bundle subsequent calls, so we need to
-	// combine the data here to avoid generating too many RPC calls.
-	mds = exportAdditionalLabels(mds)
-
-	count := 0
-	for _, md := range mds {
-		count += len(md.Metrics)
-	}
-	metrics := make([]*metricspb.Metric, 0, count)
-	for _, md := range mds {
-		if md.Resource == nil {
-			metrics = append(metrics, md.Metrics...)
-			continue
-		}
-		for _, metric := range md.Metrics {
-			if metric.Resource == nil {
-				metric.Resource = md.Resource
-			}
-			metrics = append(metrics, metric)
-		}
-	}
-	points := numPoints(metrics)
-	// The two nil args here are: node (which is ignored) and resource
-	// (which we just moved to individual metrics).
-	dropped, err := me.mexporter.PushMetricsProto(ctx, nil, nil, metrics)
-	recordPointCount(ctx, points-dropped, dropped, err)
-	return err
-}
-
-func exportAdditionalLabels(mds []*agentmetricspb.ExportMetricsServiceRequest) []*agentmetricspb.ExportMetricsServiceRequest {
-	for _, md := range mds {
-		if md.Resource == nil ||
-			md.Resource.Labels == nil ||
-			md.Node == nil ||
-			md.Node.Identifier == nil ||
-			len(md.Node.Identifier.HostName) == 0 {
-			continue
-		}
-		// MetricsToOC removes `host.name` label and writes it to node indentifier, here we reintroduce it.
-		md.Resource.Labels[conventions.AttributeHostName] = md.Node.Identifier.HostName
-	}
-	return mds
-}
-
 // pushTraces calls texporter.ExportSpan for each span in the given traces
 func (te *traceExporter) pushTraces(ctx context.Context, td pdata.Traces) error {
 	resourceSpans := td.ResourceSpans()
@@ -244,15 +116,4 @@ func (te *traceExporter) pushTraces(ctx context.Context, td pdata.Traces) error 
 	}
 
 	return te.texporter.ExportSpans(ctx, spans)
-}
-
-func numPoints(metrics []*metricspb.Metric) int {
-	numPoints := 0
-	for _, metric := range metrics {
-		tss := metric.GetTimeseries()
-		for _, ts := range tss {
-			numPoints += len(ts.GetPoints())
-		}
-	}
-	return numPoints
 }
