+++ exporter/collector/metricsexporter.go
@@ -1,0 +1,74 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// This file contains the rewritten googlecloud metrics exporter which no longer takes
+// dependency on the OpenCensus stackdriver exporter.
+
+package collector
+
+import (
+	"context"
+
+	monitoring "cloud.google.com/go/monitoring/apiv3/v2"
+	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/exporter/exporterhelper"
+	"go.opentelemetry.io/collector/model/pdata"
+)
+
+// metricsExporter is the GCM exporter that uses pdata directly
+type metricsExporter struct {
+	cfg    *Config
+	client *monitoring.MetricClient
+}
+
+func (me *metricsExporter) Shutdown(context.Context) error {
+	return me.client.Close()
+}
+
+func newGoogleCloudMetricsExporter(
+	ctx context.Context,
+	cfg *Config,
+	set component.ExporterCreateSettings,
+) (component.MetricsExporter, error) {
+	setVersionInUserAgent(cfg, set.BuildInfo.Version)
+
+	client, err := monitoring.NewMetricClient(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	mExp := &metricsExporter{
+		cfg:    cfg,
+		client: client,
+	}
+
+	return exporterhelper.NewMetricsExporter(
+		cfg,
+		set,
+		mExp.pushMetrics,
+		exporterhelper.WithShutdown(mExp.Shutdown),
+		exporterhelper.WithTimeout(exporterhelper.TimeoutSettings{Timeout: defaultTimeout}),
+		exporterhelper.WithQueue(cfg.QueueSettings),
+		exporterhelper.WithRetry(cfg.RetrySettings))
+}
+
+// pushMetrics calls pushes pdata metrics to GCM, creating metric descriptors if necessary
+func (me *metricsExporter) pushMetrics(ctx context.Context, m pdata.Metrics) error {
+	// TODO: self observability
+	points := 0
+	dropped := 0
+	var err error = nil
+	recordPointCount(ctx, points-dropped, dropped, err)
+	return nil
+}
